215. ImageBrowser (preview uploaded images)(Section 21: React Redux Application)
So now that we have upload working we will now create another component so we will be able to preview

everything that was uploaded to these blog posts so we will see all the images and from nails.

So later we'll also be able to remove those images.

So let's create another component and call it the image browser.

J.S..

So this component will be responsible for displaying all those images that we have just uploaded to

our blog posts.

So we have to import react and create a new class image browser that will extend re-act component.

As usual we render this component first.

So let's surrounded with a leaf and add some classes.

Well start with the role and the markings on top and bottom to be of the size 4 so the margin top four

and the margin bottom four for now we'll get the images from the parent component in which this would

be used.

So this would be used in the blog post form so the images will be passed as a property to this component

and then we'll iterate over every image that was passed and display it.

So we are surrounded with a diff element and will use the Greek classes so this would be we will be

using €3 for display every image that was added.

Now inside those These will put an EMG element to actually display this image.

So for now let's just use our API prefix.

Well it's not actually API prefix just it's just the prefix of our whole application which runs on the

development server for now.

And this emet would also have that EMG Floyd.

So this would be properly sized.

So at least it would be.

It can fit inside our greed.

Now we'll also have to reduce our forward the blog posts forum because that's the place where we would

like to handle all our images that were uploaded.

So if you'd like to use this image browser component in different places you can have a separate reducer

for that.

So this would be the blog post forum and what it has to have is that is sending state property and that

image which would be the currently uploaded image and the images which would be the least of all the

images.

Of course we have to add it to the reducer.

Yes to the combined reducers function to make it available inside Rydex and now inside the blog post

forum we'll add to this state to the maps state to prop's function.

So this substate the blog post form will be available here so we can just add the whole state here to

to our props.

So we'll have all the properties available that emerges emerge and is sending.

Let's go to our render method and insert this new image browser component just below that image upload

component and pass the images to it.

And first we can get this emergent property from this component props because they are mapped from the

state that we use that we map to the blog post for them.

So it says that image browser is not defined.

And this is most probably because we haven't exported that class and also haven't imported it so obviously

it can't find a definition for this class.

So now let's import statement for our image browser.

Seems we our talkin has expired so we have to log in again.

So let's do that quickly and then go back to our forum.

So this component actually gets surrendered to Bob.

Well there are no images yet and it doesn't do anything good.

So let's just leave the message nothing here yet and we'll continue in the next lecture.