216. A new reducer for BlogPostForm(Section 21: React Redux Application)
So now that we have our blog post forum and those components to upload emerge which is the image upload

and image browser will now try to make them work together.

So let's start with modifying a blog post form reduce our way of using those actions that we have added.

So first let's modify the state of the image upload request happens.

So this will only said that is sending a property of the state to true next.

That image uploaded this will only modify that is sending to false.

So that image has been uploaded so it's not sending any more.

And now we have this image uploaded which would have the action which would have the image property

added so we can send e-mail to state and to images we shoot on new items so we'll call state images

come cop which would return the array as it is right now from the state.

And will add a new element to the end of this array.

And the final action when the error happens we said that is sending two false

Now when we are uploading an image we probably shouldn't allow that form to be sent.

So let's get that is or maybe well actually we should modify this state property name from sending its

image uploading because it might be misleading because the form itself has the states submitting and

is sending doesn't really say anything about what this variable what this state property means.

So let's to image uploading and now have the button disabled when the form is submitting or when the

image is uploading.

So now we can remove this.

Nothing here yet the text from that emerged browser because there should be something there after we

upload our image.

So everything looks correct.

We have that image just passed from the blog post to form all the elements have the key because we always

have have to have a key key properties on the first element.

When we display a list.

And now let's try uploading.

So you see the bottom image is on it and it is also uploaded.

So there is a response from the server with the you are elde that we are using to display this image.

You can also see that we have this response called 201 which means that the resource was created.

And if you will choose more files you can see that the publish now button is disabled.

When the images are uploading and now after adding the fourth image you can see that also they align

pretty nicely into the columns of free images.

Maybe Let's now try to actually send a value to blog posts so them content should be at least 20 characters

and Bobby it and see what happens.

So this blog post WOZA it but it does not have any images.

You can see on the right that the images property returned from the server is empty.

And this is because we haven't sent this property with the request payload.

So we'll be working on that in the next lecture.