217.1. Clear the images when BlogPostForm unmounts(Section 21: React Redux Application)
Send the images in blogPostAdd creation action to attach the images on the blog post in the api end point.

Now let's try uploading a couple of images maybe the balloon the boat and the breach.

Now sended this new block post.

And we have two hundred one created and also the payload contains the images.

And as you see the response also has links to these images.

And if you go to the show action on the admin Bano you can set up this blog post has the free image

as a sign.

So this worked pretty well.

Now let's go back to the forum and well there is something wrong here because we send that form and

we still see the images that we have just sent.

So this is not what we would expect.

And this is normal because we only added those images to the blog post forum state which we then map

to prop's inside the blog post forum and well in that image browser too.