217.2. Clear the images when BlogPostForm unmounts(Section 21: React Redux Application)
So to clear these images from the state when we are done with creating new blog post forum we will have

to create another action which would be a blog post forum unload.

So this would be pretty simple action creator.

We don't really have to pass any data we will just create a simple event blog post form and load so

as usual we have to add a Konstanz to our Konstanz.

J.S. let's just pasted from a blog post for ok now.

Obviously we will imported.

And now we will have to use this action creator inside our map dispatched to prop's So we will be able

to dispatch this action when the form component will unload.

So this will happen for example when you actually send a form because we are doing our direct to the

main page.

We have a list of all the blog posts.

So now let's add a life cycle meathook component will unmount and inside this life cycle.

Mettled will dispatch block balls farm and laut action which will modify the state to clear every values

that we have previously added for emergents.

But first we have to of course handle that inside our reducer.

Nothing happens automatically.

Now let's modify the state accordingly.

So we leave the remaining part of the state and we are resetting the image uploading to false image

to low and the images to the empty arrays so we are just using the default values here.

Now let's try it out.

Try to upload a couple of images

and apparently we added one image but it is not being cleared.

For some reason so let's try to figure out what is wrong.

Well I see what is wrong so I've used the action block post and load and it should be block post form

unload.

We have this block post unload to clear the comments when you use when you are watching an individual

block or so.

This was only a mistake.

I just handled the wrong action inside the reducer.

So let's try it again.

Upload to images.

Go to the main page and now when you are on the forum you see there are no images.

Well there is also I think we have to handle those images are still on the server and we are not using

them so it might be a good idea to kill them but we will think about it later.