218. DELETE operation on Image (API Changes)(Section 21: React Redux Application)
Now let's think about how we can delete those images images that we have uploaded.

But maybe we've changed our mind and we do not really want them to be uploaded together with a blog

post.

So I've just restarted the server so it takes a little bit longer to finish the request.

So now I have uploaded this image and unfortunately I have no way to delete it because well we haven't

implemented anything like that.

And well actually we haven't implemented anything like that on the re-act redux application.

And also we do not really have an endpoint in our API to delete this blog posts so you can see the first

immediate problem is that they don't get deleted if you change your mind and do not post this blog post.

So let's type this command inside our API project and we will be checking the configuration that the

current configuration and also the default configuration of the Veach uploader component will be checking

one thing.

So we would like to figure out whether we'll be deleting the image file when we delete the entity.

So there is this property delete on remove that by default is set to true.

So this means that whenever we would like to delete the image entity that we are creating when we upload

an image it will also delete the file associated with it.

So we do not have to do anything extra.

So now let's go to our API project.

So we will be and may be enabling the delete operation.

So of course we have to have to delete an existing image since we are working on the items.

So this image has to exist already.

We all had an item operations and well this doesn't really require much additional configuration.

Maybe we can just check whether the user is out integrated to delete the image.

The fingers that we did not require the user to be authenticated uploaded.

We can't change it.

To require the users to be at least authenticated or maybe to also have a writer permission so they

can write a blog post.

But we haven't done that yet.

It's not a problem.

You can also add it yourself.

But well it's not a problem for us we are just learning here so let's only out the requirement to delete

the images so the user has to be Utley's right now after making this change.

We have an error in our API and it broke the whole API for us so we can't really fetch the blog posts.

So let's see what happens inside the postman.

We have an error that no item root was associated with that image resource.

So let's check the API platform documentation.

Zoom it a little bit and you see that in that item operations Dagget operation is mandatory.

So it has to be there.

So if you if you wouldn't define the item operations it would be enabled by default.

But since you did you also have to get operation because it is mandatory.

After making this change.

Well it takes a bit to refresh the API but now you can see that everything works.

And our delete operations should be available to us and we'll work on it on the next lecture to actually

delete some images.