219.1. ImageBrowser animations(Section 21: React Redux Application)
Now before will actually handle deleting our elements from the images let's first maybe animations to

our image browser so each time you match is uploaded uploaded it will be maybe more visible to the users

when this change is animated.

So the same thing we did with the command least we will have to surround everything with that transition

group and then all the elements would have to be surrounded with the CSX transition elements so lets

added as the first element that is being returned from the map of that emerges and we have to specify

some parameters.

Same way as we did with comments.

So it is the time out of the animation the class names property which will be using the CSL we defined

forwarders fade animations we can just use the same animation that we defined previously and now the

key also has to be in the first element so instead of the element inside it has to be on the scene says

transition element.

So lets see if that works.

Well apparently it works.

But now our images are very small so thats not what we expected.

And when looking at things specter of the elements you can see that we somehow got an additional empty

Deif element which we havent declared.

So this breaks our styles and it is because the transition group by default surrounds everything with

diff element.