219.2. ImageBrowser animations(Section 21: React Redux Application)
So to get rid of that and not add this element we have to specify the component property to all of that

transition group and after making that change you see that everything works as expected.

And every new image that is being added is also animated so maybe this would be more clear to the users

of the app that something has actually changed.

OK so we got that covered.

We also handled how we can fix this transition group adding an empty div element when we actually do

not want it.

Now lets add or remove bottom.

So lets copy the code from the bootstrap buttons so we will use the button that will only have an outline

so lets add this into the Deif element which is the parent element of all of those elements inside.

And also specify Margene bottom because the image already has a marking bottom so we only need a marking

bottom for our bottom.

We can also add buttons small class to make this pattern a little bit smaller because we all have lots

of them.

There will be a delete button for every possible image we upload and only change the label to remove.

Now let's test this.

And I can see that it is being added.

This Remove button is being added for every element so now will have to handle the action that clicking

this element will do something.

Now it does.

Doesn't do anything.

But this will be in the next lecture.