220. Remove button on ImageBrowser(Section 21: React Redux Application)
Last time we are removed.

But every image that we upload to the blog post.

But this button didn't function and it didn't do anything.

It was just displayed.

So now it's time that we implement the proper actions that will handle everything.

So let's start by adding action creator image delete which will accept only the ID of them image that

we would like to delete.

This will be using the phunk middle were.

So we would like to dispatch some other actions when this actually succeeds or fails.

And also when we start deleting this image so lets call a request delete.

We have this image Eurail.

Of course we do not have this metadata yet on the request So let's now add that it will accept the.

And whether this should be using the talkin to whether it should also include the token on the request.

So by default it would be true and we will send a delete request to our to the Eurail specified.

So in super agent there is a method for that and we use that token plugin plugin which accepts this

secure parameter in case we want it to require a token.

And finally we handle the response body using our methods to make it to only get the information that

we need.

Next we need to import this action creator inside the block post form and we'll pass it to the image

browser using that delete handler property.

So first let's get this from the props and parse this as a delete handler to that image browser.

Then in the image browser will get this from prop's So now we have the handler.

And this is the action to be dispatched and we need metters to handle the button click.

So let's add on click on this remove button and define metalled to that individual for every single

button click.

So we were added as a constant inside this map and call it on the image delete cleated click to accept

an event argument.

So this would be calling delete handler with the ID of the currently the current element inside the

inside the mop up method of old images.

So we'll always get the current image.

And also we called prevent a default to prevent any default behavior.

Now let's try uploading an image and pressing the remove element actually sends a request and the request

is successful.

You can see that the first is the options.

But second is delete and it returns to 100 for no content which is well the sexist response for the

delete requests because when you delete a resource you do not have any more content.

So basically the response body is empty you only get the status code.

Now let's upload another.

Imagine the balloon and it has the ID 260.

So try to remember that the last image upload it was 260.

Now click Remove and the response is success.

And now let's check if it was removed so it shouldn't be in the admin.

So it's not.

So it was successfully removed.

So our functionality is working.

And in the next electros we'll handle.

So the images that were removed will also disappear from that image browser.