222. Locking all buttons during Image upload/removal(Section 21: React Redux Application)
So now we'll handle the case that we want all the buttons to be disabled when the request is in progress

so you want to be able to remove the images like you will be able to remove more images at one time.

But first let's modify the name of the state.

Is image uploading to image request in-progress.

So we are doing it to have a generic state that we can just modify one variable and it should disable

every operation that requires backend communication with the API.

We also have to modify the name of this inside block post form and inside that disabled property set

to the bottom so also when we have this image request in progress we should be able to pass this somehow

to our image browser because it is not connected to this state.

So let's add a new prop on this on this image browser cold image.

Sorry cold is locked and we'll pass that request in progress to this is locked up.

And now inside the image browser will have disabled property to all of the buttons that we had for every

uploaded image.

So first let's get this to the variable East looked from from the props.

And this button will be disabled when and any request is in progress.

So first to our image upload is actually starting and image upload request but also we will need another

action

and this would be called Image delete request.

So let's do this auction creator and also a new constant for this action image delete request and this

auction will be dispatched before we actually make our request that that is the link that emerge from

the list also mind that you have to import this image delete request Constand inside this file.

So before we actually make a request we will dispatch the image delete request.

So we'll have the same property modified whether we are uploading or we are deleting that image.

So this looks like we can use a couple of those state modifiers the same way.

So for example when we do an image upload request we said that image request in progress as to true.

So we can do the same for image delete requests or we do not have to repeat ourselves because we also

want it to be modified to true.

Now let's also look where we should change this state.

So this is probably this also should probably happen inside the image deleted because we said previously

the state to that image requesting progress when we started getting the image.

So when it is done we should now set this to fall.

So we have now the common state for uploading for deleting and this will disable all of that but all

the buttons when either the upload is in progress or removing of an image is in progress.

So let's upload a couple of these images.

And now try removing one so you can see that before the request finishes.

It disables all the buttons including the remove buttons on all the other images and also the poppys

now button.

So you shouldn't be you wouldn't be able to send new blog post with an image that was already deleted.

And also when you are uploading new images you wouldn't be able to delete the images that were previously

uploaded.